+++ packages/plugin-truffle/src/index.ts
@@ -6,6 +6,7 @@ export * from './upgrade-proxy';
 export { deployBeacon } from './deploy-beacon';
 export { deployBeaconProxy } from './deploy-beacon-proxy';
 export { upgradeBeacon } from './upgrade-beacon';
+export { forceImport } from './force-import';
 export { admin } from './admin';
 export { erc1967 } from './erc1967';
 export { beacon } from './beacon';